requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.14
Django==3.0.2
django-bootstrap4==2.3.1
django-cleanup==5.1.0
gensim==3.8.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==2.1.1
isort==5.6.4
Janome==0.4.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.3
pandas==1.1.5
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
rsa==4.7
scipy==1.5.4
six==1.15.0
smart-open==4.1.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.26.2
wordcloud==1.8.1
wrapt==1.12.1
zipp==3.4.0
